First Cut at a ZIO like Schedule.
The environment does not integrate correctly so will stop here and pivot.

The general concept is that you have a function from schedule status (current iteration state of the schedule) returning a schedule result. (Stop, run again/run again with delay). And you fold over the schedule.

That part is ok, its not as unified as the ZIO one, but is better than the Haskell one.

Performance will likely be an issue with all the closures required to keep the types happy. The need to eliminate some the generics is what makes it hard. (backwards compatible with the current static schedule)

Might take some stuff from this for the next attempt.

I think the schedule will need to be specialized 4 times to make environment work,
2 for Eff and 2 for Aff. Then a generic free version that works over both. (That one might be able to extend from the Eff<A> variant and set A to Unit)

I also want to include in it the ability to thread some user defined state S through it like the ZIO one. This makes it possible to remove stuff from the schedule status. That way the schedule is responsible for defining what it requires on each step.